requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
numpy==1.19.5
openpyxl==3.0.6
pipenv==2020.11.15
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
virtualenv==20.4.0
virtualenv-clone==0.5.4
Werkzeug==1.0.1
xlrd==2.0.1
zope.interface==5.2.0
